requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
faiss-cpu
groq
langchain-groq
sentence_transformers
ragas 
datasets
